backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25

# OpenAI and AI
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5

# Data Processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
Pillow==10.2.0

# Document Export
weasyprint==61.2
pydyf==0.10.0
markdown==3.10.2
markdown2==2.4.12

# Vector Search
faiss-cpu==1.7.4

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0
beautifulsoup4==4.12.3

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
hypothesis==6.98.3
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
